# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted electronics parts inventory. Track quantity, package, brand, location, and vendor pricing across your parts, scan a distributor barcode to enrich a part, and generate bin and part labels. All protected endpoints require a Bearer JWT (or `fbin_pat_*` personal access token) in the Authorization header.",
        "title": "FireBin API",
        "contact": {
            "name": "FireBin",
            "url": "https://github.com/FireBall1725/firebin"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "26.7.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return all categories.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "List categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.Category"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new category, optionally nested under a parent.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Create category",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.Category"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a category that has no parts assigned to it.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Delete category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing category's name, parent, or description.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Update category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.Category"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/locations/{id}/stock": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return the stock contents of a bin (scan-a-bin).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "List location stock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.StockItem"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/parts/low-stock": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return parts at or below their minimum stock level.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "List low-stock parts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.Part"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return the dashboard summary counts and totals.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get dashboard stats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.Stats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/stock/recent": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return the newest stock movements across all parts.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "List recent stock activity",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.StockTransaction"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_firelabsca_firebin-api_internal_models.Category": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "part_count": {
                    "description": "parts directly in this category (List only)",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_firelabsca_firebin-api_internal_models.ManufacturerPart": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "datasheet_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "manufacturer_id": {
                    "type": "string"
                },
                "manufacturer_name": {
                    "type": "string"
                },
                "mpn": {
                    "type": "string"
                },
                "part_id": {
                    "type": "string"
                },
                "supplier_parts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.SupplierPart"
                    }
                }
            }
        },
...
        "github_com_firelabsca_firebin-api_internal_models.Part": {
            "type": "object",
            "properties": {
                "alternatives": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.PartAlternative"
                    }
                },
                "barcode": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "default_location_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_path": {
                    "type": "string"
                },
                "ipn": {
                    "description": "FireBin internal part number",
                    "type": "string"
                },
                "is_assembly": {
                    "type": "boolean"
                },
                "is_component": {
                    "type": "boolean"
                },
                "is_purchaseable": {
                    "type": "boolean"
                },
                "is_template": {
                    "type": "boolean"
                },
                "is_trackable": {
                    "type": "boolean"
                },
                "keywords": {
                    "type": "string"
                },
                "manufacturer_parts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.ManufacturerPart"
                    }
                },
                "minimum_stock": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "package": {
                    "type": "string"
                },
                "parameters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.PartParameter"
                    }
                },
                "primary_location": {
                    "type": "string"
                },
                "primary_location_id": {
                    "type": "string"
                },
                "primary_manufacturer": {
                    "type": "string"
                },
                "primary_mpn": {
                    "description": "Primary MPN/manufacturer and bin, for the parts list (from List's laterals).",
                    "type": "string"
                },
                "total_stock": {
                    "description": "Joined / computed, populated on some reads.",
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "variant_count": {
                    "type": "integer"
                },
                "variant_of": {
                    "type": "string"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.Part"
                    }
                }
            }
        },
        "github_com_firelabsca_firebin-api_internal_models.PartAlternative": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "manufacturer": {
                    "type": "string"
                },
                "mpn": {
                    "type": "string"
                },
                "part_id": {
                    "description": "set if we stock it",
                    "type": "string"
                },
                "part_name": {
                    "type": "string"
                }
            }
        },
        "github_com_firelabsca_firebin-api_internal_models.PartParameter": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "template_name": {
                    "type": "string"
                },
                "units": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "github_com_firelabsca_firebin-api_internal_models.PriceBreak": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "number"
                }
            }
        },
...
        "github_com_firelabsca_firebin-api_internal_models.Stats": {
            "type": "object",
            "properties": {
                "inventory_value": {
                    "type": "number"
                },
                "locations_count": {
                    "type": "integer"
                },
                "low_stock_count": {
                    "type": "integer"
                },
                "parts_count": {
                    "type": "integer"
                },
                "total_units": {
                    "type": "number"
                },
                "variants_count": {
                    "type": "integer"
                }
            }
        },
        "github_com_firelabsca_firebin-api_internal_models.StockItem": {
            "type": "object",
            "properties": {
                "added_at": {
                    "type": "string"
                },
                "barcode": {
                    "description": "scannable lot identity (a mini spool)",
                    "type": "string"
                },
                "batch": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_path": {
                    "type": "string"
                },
                "location_id": {
                    "type": "string"
                },
                "location_name": {
                    "type": "string"
                },
                "name": {
                    "description": "human label for the lot (\"Mini spool #1\")",
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "part_id": {
                    "type": "string"
                },
                "part_name": {
                    "type": "string"
                },
                "purchase_price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "number"
                },
                "serial": {
                    "type": "string"
                },
                "split_from": {
                    "description": "the lot this was cut from",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "supplier_part_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_firelabsca_firebin-api_internal_models.StockTransaction": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "delta": {
                    "type": "number"
                },
                "from_location_id": {
                    "type": "string"
                },
                "from_location_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "part_id": {
                    "type": "string"
                },
                "part_name": {
                    "type": "string"
                },
                "resulting_quantity": {
                    "type": "number"
                },
                "stock_item_id": {
                    "type": "string"
                },
                "to_location_id": {
                    "type": "string"
                },
                "to_location_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_firelabsca_firebin-api_internal_models.SupplierPart": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "manufacturer_part_id": {
                    "type": "string"
                },
                "moq": {
                    "type": "number"
                },
                "packaging": {
                    "type": "string"
                },
                "pricing": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.PriceBreak"
                    }
                },
                "sku": {
                    "type": "string"
                },
                "supplier_id": {
                    "type": "string"
                },
                "supplier_name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}